The Concurrent_Observer is a concurrent variant of the Publisher/Subscriber (Observed/Observer in EPOS) GoF design pattern.
The main difference from the traditional Publisher/Subscriber is that the notify method of the publisher does not invoke the update method of its subscribers. Instead, the publisher  only notifies its subscribers that a change has occurred. The subscribers, usually running on another thread, wait for such change notification and then perform their update.

In the version 2 of Concurrent_Observer the subscriber can be either a Concurrent_Observer object or a Handler. The publisher (Concurrent_Observed) has a list of handlers as its subscribers. Such handlers can belong to an Concurrent_Observer or not (i.e. the use of the Concurrent_Observer is entirely optional).

The Concurrent_Observed::notify method invokes, for each handler, its call operator. The semantics of such operator is the one defined by the subclasses of Handler (e.g. Semaphore_Handler).

The Concurrent_Observer::update method invokes the "dual call operator" of its handler. For example, for a Semaphore_Handler its call operator is the v() operation of Semaphore and its dual call operator is the p() operation. From the point of view of the Concurrent_Observer pattern, the call operator is the one used to **notify a change** (used by the publisher) and the dual call operator is the one to **wait for a change** (used by the subscriber).

Details:
* Added a _link attribute to Handler.
* Created Dual_Handler class (subclass of Handler) and made Semaphore_Handler, Mutex_Handler, and Condition_Handler extended it.



git-svn-id: https://epos.lisha.ufsc.br/svn/epos2@3450 86035c5f-d46c-4136-a2d5-fe63f1d4a782