Hyper EPOS Milestone 1 (HE 1)
* HE 1 - Load Tasks from Distinct Programs

Description:
Implement exec-like function for EPOS, enabling it to load a distinct program (ELF file) for each task.

Application: Main Thread / Main Task (Domain0) creates a Task A and a Thread A that is load from an ELF file and prints "a"s on the screen.
Main Thread also creates Task B and Thread B (also loaded from an ELF file) and prints "b"'s on the screen.

Status: Completed with issues (see ISSUES #1 to #3 at the code comments).



git-svn-id: https://epos.lisha.ufsc.br/svn/epos2@3451 86035c5f-d46c-4136-a2d5-fe63f1d4a782