* Refactoring E100 so it supports QEMU and physical device. i82559ER is the engine that shall be used for QEMU tests.
Regarding the transmission:
QEMU 2.4 requires from the TCB to use a TBD even while using simplified
memory structure.
Because of that, we keep tbd_array = 0xFFFFFFFF and tbd_number = 0
Such workaround is not expected to work on a physical E100.
For that, prefer using i82559c instead.
In order to unify i82559ER and i82559c regarding transmission, one can program
E100 to use the flexible memory mode (instead of the simplified one) always
using TBD.

Regarding the reception:
There seems to be a bug on QEMU E100 device: it does not update the actual_count
field of RFD. A patch for QEMU is proposed and a workaround for the hardware
mediator is presented.

git-svn-id: https://epos.lisha.ufsc.br/svn/epos2@4321 86035c5f-d46c-4136-a2d5-fe63f1d4a782